Add dna sprite
Aminos need to be stored on a parent dna sprite so we have an easier
time referring to them (for getting last position etc).

Aminos are stored in a queue to make it easier to get rid of the old
ones in the future.

Should the borders also live on the dna sprite?